Add files via upload
Hogwarts Legacy is a game about Harry Potter Universe released in 2023 made by collaboration of Warner Bros, Avalanche Software and Portkey Games. It showcases the life and roleplay of students in Hogwarts - School of Witchcraft and Wizardry in 1800. The player controls a character going through challenges and missions while attending the school and making various friendships with many characters.

In this analysis, datatset of reviews of the game were used to showcase the opinion of users in the last two years. This dataset is derived from kaggle and will showcase the positive, negative and neutral reviews using 3 Machine Learning models - Logistic Regression, SVM and Random Forest Classifier. After that, we hypertuned models to have the best output possible from dataset.